requirements.txt
opencv-python
numpy
requests
urllib3
ultralytics
python-dotenv
